feat: Refactor clean_up fixture to be able to use in localstack-ext
